Use default value of T in ExecuteScalar when column type is Null
When executing a SUM query the method throws NullReferenceException if the
table is empty or no rows were selected by the WHERE clause or similar.
In these cases SQLite3.ColumnType returns ColType.Null, ReadCol returns null,
and the attempted cast to T throws the exception.

This fixes koush/sqlite-net#171.

Conflicts:
	src/SQLite.cs

Merged-by: Øystein Krog <oystein.krog@gmail.com>